examples/.template/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools_scm]
git_describe_command = "git describe --long --first-parent"
root = "../../.."

[project]
name = "nat_EXAMPLE_NAME"
dynamic = ["version"]
dependencies = ["nvidia-nat[langchain]~=1.3"]
requires-python = ">=3.11,<3.13"
description = "EXAMPLE_DESCRIPTION - NeMo Agent Toolkit example"
keywords = ["ai", "rag", "agents"]
classifiers = ["Programming Language :: Python"]

[tool.uv.sources]
nvidia-nat = { path = "../../..", editable = true }

[project.entry-points.'nat.components']
nat_EXAMPLE_NAME = "nat_EXAMPLE_NAME.register"
